fix(money): work_report / technical_inspection en Decimal — dette ADR-0008 videe
Suite de #661. Le gate `scripts/check-no-f64-money.sh` avait GELE une liste de
`f64` monetaires sous la rubrique « DETTE CONNUE ET TRACEE (a resorber, pas un
carve-out accorde) ». `payment_reminder` a ete resorbe le 2026-08-26 ; ce lot
resorbe TOUT LE RESTE. La rubrique est desormais VIDE.

Ancrage metier : ces couts sont refactures aux coproprietaires via la
repartition des charges (Art. 3.86 CC) et un cout de travaux entre dans le
calcul du fonds de reserve. Ce ne sont pas des valeurs d'affichage.

Trois defauts de nature differente, pas un seul :

1. DEUX COLONNES EN DOUBLE PRECISION. `work_reports.cost` (NOT NULL) et
   `technical_inspections.cost` (nullable), de la colonne SQL au DTO. Migration
   20260829000000, NUMERIC(12,2) aligne sur `expenses.amount`.

   `ALTER COLUMN TYPE` NE REECRIT PAS l'expression d'un CHECK : il y insere un
   cast vers l'ancien type. Reproduit sur une base reelle AVANT d'ecrire le
   fichier — `CHECK (cost >= 0)` devient
   `CHECK ((cost)::double precision >= (0)::double precision)`. La positivite
   serait restee evaluee en binary64 APRES une migration dont c'est l'objet
   meme. Les CHECK sont donc droppes puis recrees avec des litteraux NUMERIC.

   Les noms de contraintes ne sont pas devines — le defaut n.1 de la migration
   `payment_reminders` venait precisement de noms inventes. Releves dans
   `pg_constraint` sur une base reconstruite depuis le DDL d'origine. Ce sont
   ici des contraintes de COLONNE, donc nommees de facon deterministe,
   contrairement aux CHECK de TABLE anonymes de `payment_reminders`.

2. UN ALLER-RETOUR Decimal -> f64 GRATUIT SUR LE TABLEAU DE BORD.
   `stats_repository_impl` degradait `expenses.amount` — deja NUMERIC(12,2)
   depuis 20260502000000 — par `SUM(amount)::float8`, a deux endroits. Meme
   defaut que celui trouve dans `find_overdue_expenses_without_reminders` sous
   #661. Casts retires.

3. UN INVARIANT QUI ALLAIT DISPARAITRE SILENCIEUSEMENT. `validator` ne sait pas
   borner un `Decimal` : `#[validate(range(min = 0.0))]` faisait echouer la
   compilation. Plutot que de retirer l'annotation (et l'invariant avec), la
   regle descend dans le domaine — erreurs typees `WorkReportError` /
   `TechnicalInspectionError` mappees en `AppError::Validation` (400, jamais
   500). Effet de bord favorable : le chemin de mise a jour ecrivait
   `work_report.cost = cost` directement, l'invariant du constructeur ne s'y
   appliquait pas ; il couvre desormais LES DEUX points d'ecriture.

CE QUE LA SEULE CONVERSION DE TYPE AURAIT CASSE — deux regressions ecartees
apres verification empirique, pas par raisonnement :

  f64 (avant)                     -> 1500.0
  Decimal nu                      -> "1500.00"   <- CHAINE, casse le contrat
  #[serde(with = "...::float")]   -> 1500.0

  Et `#[serde(with = ...)]` fait perdre le defaut implicite d'un `Option`
  absent : sans `#[serde(default)]`, un `cost` omis dans une mise a jour
  partielle serait passe de « champ non modifie » a 400.

TESTS : 22 tests 4-cat sur les deux entites. Les @edge ne verifient pas que la
borne, ils verifient `0.10 + 0.20 == 0.30`, egalite FAUSSE en binary64 — le
test porte sur la raison d'etre de la conversion, pas sur son type. Les
@negative verifient qu'un refus ne laisse aucune ecriture partielle,
`updated_at` compris. Les @security ancrent le pourquoi : un cout negatif
refacture via la repartition des charges produirait un AVOIR au profit des
coproprietaires depuis un simple rapport de travaux.

VERIFIE, tout execute pour de vrai :
  cargo test --lib                          1684 passed, 0 failed, 8 ignored
  cargo clippy --all-targets --all-features -- -D warnings   exit 0
  cargo run --bin export_openapi + diff     aucun drift
  cargo check --all-targets                 exit 0, aucun avertissement
  cargo fmt --check                         propre
  scripts/check-no-f64-money.sh             vert AVEC l'allowlist videe
  migration rejouee sur base neuve          contraintes en numeric, sans cast
                                            residuel ni doublon, donnees intactes

DEUX DEFAUTS TROUVES AU PASSAGE, TRACES ET NON CORRIGES (elargir en douce
irait contre la discipline de perimetre du depot) :

  - `WorkReportFilters` expose 9 champs, `work_report_repository_impl` n'en
    applique que 2. `min_cost`, `max_cost`, `warranty_type`, `contractor_name`,
    `work_date_from`, `work_date_to`, `warranty_active` sont acceptes par l'API
    puis SILENCIEUSEMENT IGNORES. Un filtre qui ment est pire qu'un filtre
    absent, parce que rien ne le signale.
  - `InspectionList.svelte:78` / `InspectionDetail.svelte:98` envoient
    `cost: form.cost || undefined` — `0` etant falsy, un cout de zero devient
    « champ non fourni », alors que le domaine l'accepte explicitement.

Journal : docs/agent-activity/2026-08-29-work-report-inspection-decimal.md

Refs: #661

+++ backend/src/application/dto/filters.rs
@@ -80,8 +80,18 @@ pub struct WorkReportFilters {
     pub contractor_name: Option<String>,
     pub work_date_from: Option<DateTime<Utc>>,
     pub work_date_to: Option<DateTime<Utc>>,
-    pub min_cost: Option<f64>,
-    pub max_cost: Option<f64>,
+    // ADR-0008 : bornes de coût en `Decimal`, comme la colonne
+    // `work_reports.cost` qu'elles filtrent.
+    //
+    // ATTENTION — ces deux bornes, comme `warranty_type`, `contractor_name`,
+    // `work_date_from`, `work_date_to` et `warranty_active`, sont acceptées
+    // par l'API mais **jamais appliquées** : `work_report_repository_impl`
+    // ne lit que `building_id` et `work_type`. Un appelant qui passe
+    // `?min_cost=1000` reçoit la liste NON filtrée en croyant l'inverse.
+    // Défaut constaté en convertissant ces champs — tracé, non corrigé ici
+    // (le corriger demande 7 filtres + tests 4-cat, hors périmètre ADR-0008).
+    pub min_cost: Option<Decimal>,
+    pub max_cost: Option<Decimal>,
     pub warranty_active: Option<bool>,
 }
